feat: Restructure agent-cloud with proper routing and fixes
Routing Structure:
- /agent-cloud → Redirects to /agent-cloud/new
- /agent-cloud/new → New session page with repo selector and chat input
- /agent-cloud/session?id=xxx → Active session page

Layout Changes:
- Remove chat input from sidebar (sessions list only)
- Add shared layout with context for state management
- Sidebar width reduced to 288px (cleaner)

New Session Page:
- Full chat input with repo/branch/model selectors
- Search repos in dropdown
- Auto-resize textarea (max 200px)

Session Page:
- GitHub token passed to AI in prompts for auth operations
- Chat input with reduced height (44px min, 120px max, then scrolls)
- Markdown rendering with Streamdown Response component

Fixes:
- Removed typing lag by using native textarea without complex state
- Proper auto-resize with max height limit
- Scroll within textarea when content exceeds max height
